requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.29
botocore==1.34.29
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1.7
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-js-asset==2.2.0
django-multiupload==0.6.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.6.0
django-storages==1.14.2
idna==3.6
jmespath==1.0.1
phonenumbers==8.13.29
pillow==10.2.0
psutil==5.9.8
PyMySQL==1.1.0
python-dateutil==2.8.2
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
